jenkinsAutomation/new_requirements.txt
Django==3.1.12
django-cors-headers==3.7.0
django-filter==21.1
django-nacl-fields==4.0.0
django-redis==5.2.0
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.12.4